% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.bpmodel}
\alias{plot.bpmodel}
\title{View all or a subset of breakpoint models}
\usage{
\method{plot}{bpmodel}(model, ref = NULL, dir = NULL,
  only.calls = FALSE)
}
\arguments{
\item{model}{S3 object supplied by running proteoformR}

\item{ref}{A vector of references to build plots for. The list of references (only.calls == FALSE) or only those with breakpoint calls (only.calls == TRUE) 
will be passed unfiltered if ref is set to "All", and will be filtered for only the first if ref is NULL.}

\item{dir}{Directory to save plot if supplied, but NULL specifies print.}

\item{only.calls}{Whether to plot only references with breakpoints called if only.calls == TRUE.}
}
\description{
\code{plot.bpmodel} builds ggplot2 breakpoint figures for criticism. 
The default behavior is to build figures for all references that proteoformR called breakpoints and save them to the working directory. 
Uses a color safe pallate for now, which can plot up to 8 samples.
}
\seealso{
\code{\link{proteoformR}}
}
