% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.bpmodel.single}
\alias{plot.bpmodel.single}
\title{View a single breakpoint model}
\usage{
\method{plot}{bpmodel.single}(model, ref = NULL, dir = NULL)
}
\arguments{
\item{model}{S3 object supplied by running proteoformR}

\item{ref}{The protein reference to extract and plot}

\item{dir}{Directory to save plot if supplied, but NULL specifies print}
}
\description{
\code{plot.bpmodel.single} builds single ggplot2 breakpoint figures for criticism. Can be called specifically,
but was originally intended to be called from plot.bpmodel. The default behavior, by leaving
ref and dir null, is to plot the first reference and print to stdout. Uses a color safe pallate
for now, which can plot up to 8 samples.
}
\seealso{
\code{\link{proteoformR}} and \code{\link{plot.bpmodel}}
}
