\name{proteoformR-package}
\alias{proteoformR-package}
\alias{proteoformR}
\docType{package}
\title{
  Identify changepoints in peptide properties
}
\description{
  Solves an L0 penalized changepoint model for each protein accross whole proteome mass spectrometry datasets.
  Homogenizes the modeling process by Z-scoring measured values, thus allowing this tool to be used with theoretically
  any continuously valued measurement. Includes ability to specify batch values, such as sample, to increase power.
}
\details{
  Coming soon
}
\author{
Author, Maintainer: Anthony Valente <valenta4@uw.edu>
}
\references{
  Coming soon
}
\keyword{ package }
