% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.bpmodel}
\alias{plot.bpmodel}
\title{View all or a subset of breakpoint models}
\usage{
\method{plot}{bpmodel}(model, ref = "All", dir = getwd(),
  only.calls = TRUE)
}
\arguments{
\item{model}{S3 object supplied by running proteoformR}

\item{ref}{A vector of references to build plots for. If the default value is used, "All",
then plots will be generated for references with breakpoint calls (if only.calls == TRUE), or all references in the set (if only.calls == TRUE).}

\item{dir}{Directory to save plot if supplied, but NULL specifies print. Defaults to the working directory.}

\item{only.calls}{Whether to plot only references with breakpoints called if only.calls == TRUE}
}
\description{
\code{plot.bpmodel} builds ggplot2 breakpoint figures for criticism. 
The default behavior is to build figures for all references that proteoformR called breakpoints and save them to the working directory. 
Uses a color safe pallate for now, which can plot up to 8 samples.
}
\seealso{
\code{\link{proteoformR}}
}
