% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteoformr.R
\name{proteoformR}
\alias{proteoformR}
\title{Build a breakpoint model for each reference in a tidy format dataframe.}
\usage{
proteoformR(data, vals, ref = NULL, start = NULL, end = NULL,
  batch = NULL, lambda = "bic", model_type = "reference")
}
\arguments{
\item{data}{A dataframe in tidy format containing at least one column of continuous values}

\item{vals}{Column name for values within data}

\item{ref}{Column name for strings which partition values into protein references}

\item{start}{Start position for each peptide}

\item{end}{Ending position for each peptide}

\item{batch}{Batching information. In most cases this will be the sample number for each measurement.}

\item{lambda}{Penalty term to apply to each L0 optimization. Must be one of 'bic', 'aic', or a numeric value}

\item{model_type}{Specify whether model shares mean shifts between replicates or allows each replicate
to have its own mean shift while maintaining common breakpoints.}
}
\value{
An S3 object with class of type, "bpmodel", which contains the manipulated input, 
a dataframe of breakpoints, and a predicted best model.
}
\description{
\code{proteoformR} uses a Rcpp dynamic programming backend to solve an L0 penalized changepoint model for each protein. 
For the best performance, each sample should be treated as a batch, but the only required arguments are data and vals; 
all other arguments are suggestions and will only be used if supplied. This function performs sorting and scaling internally.
}
